Register only the timestamp of the last interaction
This is only used to determine if the session has expired, so no more data is necessary.
This also lets us rename the confusing lastChangedInteraction() method to the simpler lastInteraction().

Also moved all public methods above the private methods,
so it's easier to get a feel for the public interface of this class